requirements.txt
# ─────────────────────────────────────────────
#  Digital Twin – Python Dependencies
# ─────────────────────────────────────────────

# LangChain core framework
langchain==0.3.25
langchain-core==0.3.58
langchain-community==0.3.23

# Groq LLM integration (fast & free)
langchain-groq==0.3.2

# HuggingFace embeddings (free, runs locally)
langchain-huggingface==0.1.2
sentence-transformers==4.1.0

# Vector database (runs locally, no server needed)
faiss-cpu==1.10.0

# PDF reading
pypdf==5.4.0

# Tools
duckduckgo-search==7.5.3       # Free web search, no API key needed
wikipedia==1.4.0               # Wikipedia lookup

# Streamlit for the chat UI
streamlit==1.45.0

# Environment variable management
python-dotenv==1.1.0

# Utilities
numexpr==2.10.2                # Fast math expressions for calculator
requests==2.32.3